# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: New Face Global Network API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create product (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /products/slug/{slug}:
    get:
      operationId: getProductBySlug
      tags: [products]
      summary: Get product by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
        "404":
          description: Not found

  # CART
...
  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get app settings
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppSettings"
    put:
      operationId: updateSettings
      tags: [settings]
      summary: Update app settings (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppSettings"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppSettings"

  # PROMO CODES
  /promos:
    get:
      operationId: listPromos
      tags: [promos]
      summary: List promo codes (admin)
      responses:
        "200":
          description: Promos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PromoCode"
    post:
      operationId: createPromo
      tags: [promos]
      summary: Create promo code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePromoBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromoCode"
  /promos/validate:
    post:
      operationId: validatePromo
      tags: [promos]
      summary: Validate promo code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidatePromoBody"
      responses:
        "200":
          description: Promo valid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromoCode"
        "404":
          description: Invalid code

  # REPORTS & DASHBOARD
...
components:
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        price:
          type: number
        comparePrice:
          type: ["number", "null"]
        image:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        stock:
          type: integer
        featured:
          type: boolean
        isProPackage:
          type: boolean
        status:
          type: string
        commissionRate:
          type: number
        createdAt:
          type: string
      required:
        - id
        - name
        - slug
        - description
        - price
        - stock
        - featured
        - isProPackage
        - status
        - commissionRate
        - createdAt
    ProductDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        price:
          type: number
        comparePrice:
          type: ["number", "null"]
        images:
          type: array
          items:
            type: string
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        stock:
          type: integer
        featured:
          type: boolean
        isProPackage:
          type: boolean
        status:
          type: string
        commissionRate:
          type: number
        ingredients:
          type: ["string", "null"]
        benefits:
          type: ["string", "null"]
        relatedProducts:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        createdAt:
          type: string
      required:
        - id
        - name
        - slug
        - description
        - price
        - images
        - stock
        - featured
        - isProPackage
        - status
        - commissionRate
        - relatedProducts
        - createdAt
    ProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - products
        - total
        - page
        - totalPages
    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        price:
          type: number
        comparePrice:
          type: ["number", "null"]
        image:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        stock:
          type: integer
        featured:
          type: boolean
        isProPackage:
          type: boolean
        commissionRate:
          type: number
        ingredients:
          type: ["string", "null"]
        benefits:
          type: ["string", "null"]
      required:
        - name
        - slug
        - description
        - price
        - stock
        - featured
        - isProPackage
        - commissionRate
...
    AppSettings:
      type: object
      properties:
        companyName:
          type: string
        companyLogo:
          type: ["string", "null"]
        contactEmail:
          type: string
        contactPhone:
          type: ["string", "null"]
        taxRate:
          type: number
        shippingRate:
          type: number
        freeShippingThreshold:
          type: number
        paymentMethods:
          type: array
          items:
            type: string
        cashAppHandle:
          type: ["string", "null"]
        paypalEmail:
          type: ["string", "null"]
        registrationPackagePrice:
          type: number
        registrationPackageId:
          type: ["integer", "null"]
        homePageBanner:
          type: ["string", "null"]
        homePageBannerSubtitle:
          type: ["string", "null"]
        demoMode:
          type: boolean
      required:
        - companyName
        - contactEmail
        - taxRate
        - shippingRate
        - freeShippingThreshold
        - paymentMethods
        - registrationPackagePrice
        - demoMode
    PromoCode:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        discountType:
          type: string
        discountValue:
          type: number
        minOrderAmount:
          type: number
        maxUses:
          type: ["integer", "null"]
        usedCount:
          type: integer
        expiresAt:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - code
        - discountType
        - discountValue
        - minOrderAmount
        - usedCount
        - isActive
        - createdAt
    CreatePromoBody:
      type: object
      properties:
        code:
          type: string
        discountType:
          type: string
        discountValue:
          type: number
        minOrderAmount:
          type: number
        maxUses:
          type: ["integer", "null"]
        expiresAt:
          type: ["string", "null"]
      required:
        - code
        - discountType
        - discountValue
        - minOrderAmount
    ValidatePromoBody:
      type: object
      properties:
        code:
          type: string
        orderAmount:
          type: number
      required:
        - code
        - orderAmount
...